+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Release 1.13.0
 ## Arena
 - Added More Slugcat's "Challenges" to Arena! 
-
+- Synced round kills
+- ⚠️ BREAKING: Updated  `GetPlayerTrophies ` to  `GetAllPlayerTrophies ` and  `GetRoundPlayerTrophies`
 ## Story
 - Fixed the "Wait for others to rescue you" death prompt blocking pause inputs.